pyproject.toml
[project]
name = "health-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "icecream==2.1.3",
    "notion-client==2.7.0",
    "pandas==2.3.2",
    "plotly==5.18.0",
    "python-dotenv==1.1.1",
    "streamlit==1.38.0",
    "libsql-client==0.3.0",
    "libsql==0.1.11",
    "pyturso==0.4.2",
    "aiosqlite>=0.22.1",
]
